Add more useful error when attempting to 'show' files in components/
If an env isn't supplied with the kubecfg show -f <FILE> command, where
the file is present in the components/ directory, kubecfg can't load the
appropriate ksonnet lib files from /environments with the -J flag.

This commit adds an INFO message on `show` failures related to expanding
the jsonnet file that provides the user with more context.